feat(worldgen): add carveBuildingRect helper and delegate hollow-rectangle carving to it in town_gen
